+++ src/main/java/com/direwolf20/buildinggadgets/common/items/gadgets/GadgetDestruction.java
@@ -84,6 +84,7 @@ public class GadgetDestruction extends GadgetGeneric {
         if (SyncedConfig.nonFuzzyEnabledDestruction)
             list.add(TextFormatting.GOLD + I18n.format("tooltip.gadget.fuzzy") + ": " + getFuzzy(stack));
 
+        addInformationRayTraceFluid(list, stack);
         addEnergyInformation(list,stack);
     }
 
@@ -213,7 +214,7 @@ public class GadgetDestruction extends GadgetGeneric {
         player.setActiveHand(hand);
         if (!world.isRemote) {
             if (!player.isSneaking()) {
-                RayTraceResult lookingAt = VectorTools.getLookingAt(player);
+                RayTraceResult lookingAt = VectorTools.getLookingAt(player, stack);
                 if (lookingAt == null && getAnchor(stack) == null) { //If we aren't looking at anything, exit
                     return new ActionResult<ItemStack>(EnumActionResult.FAIL, stack);
                 }
@@ -238,7 +239,7 @@ public class GadgetDestruction extends GadgetGeneric {
     public static void anchorBlocks(EntityPlayer player, ItemStack stack) {
         BlockPos currentAnchor = getAnchor(stack);
         if (currentAnchor == null) {
-            RayTraceResult lookingAt = VectorTools.getLookingAt(player);
+            RayTraceResult lookingAt = VectorTools.getLookingAt(player, stack);
             if (lookingAt == null) {
                 return;
             }