+++ src/main/java/com/direwolf20/buildinggadgets/common/tools/BuildingModes.java
@@ -46,6 +46,11 @@ public class BuildingModes {
         List<BlockPos> coordinates = new ArrayList<BlockPos>();
         BlockPos playerPos = new BlockPos(Math.floor(player.posX), Math.floor(player.posY), Math.floor(player.posZ));
         BlockPos pos = startBlock;
+        boolean onTop = GadgetBuilding.shouldPlaceAtop(tool);
+        int offset = onTop ? 1 : 0;
+        if (mode != GadgetBuilding.ToolMode.BuildToMe && mode != GadgetBuilding.ToolMode.Grid && mode != GadgetBuilding.ToolMode.Surface)
+            range -= 1 - offset;
+
         int bound = (range - 1) / 2;
         EnumFacing playerFacing = player.getHorizontalFacing();
         int boundX, boundZ;
@@ -61,42 +66,42 @@ public class BuildingModes {
         //***************************************************
         if (mode == GadgetBuilding.ToolMode.BuildToMe) {
             if (sideHit == EnumFacing.SOUTH) {
-                for (int i = startBlock.getZ() + 1; i <= playerPos.getZ() - 1; i++) {
+                for (int i = startBlock.getZ() + offset; i <= playerPos.getZ() - 1; i++) {
                     pos = new BlockPos(startBlock.getX(), startBlock.getY(), i);
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.NORTH) {
-                for (int i = startBlock.getZ() - 1; i >= playerPos.getZ() + 1; i--) {
+                for (int i = startBlock.getZ() - offset; i >= playerPos.getZ() + 1; i--) {
                     pos = new BlockPos(startBlock.getX(), startBlock.getY(), i);
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.EAST) {
-                for (int i = startBlock.getX() + 1; i <= playerPos.getX() - 1; i++) {
+                for (int i = startBlock.getX() + offset; i <= playerPos.getX() - 1; i++) {
                     pos = new BlockPos(i, startBlock.getY(), startBlock.getZ());
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.WEST) {
-                for (int i = startBlock.getX() - 1; i >= playerPos.getX() + 1; i--) {
+                for (int i = startBlock.getX() - offset; i >= playerPos.getX() + 1; i--) {
                     pos = new BlockPos(i, startBlock.getY(), startBlock.getZ());
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.UP) {
-                for (int i = startBlock.getY() + 1; i <= playerPos.getY() - 1; i++) {
+                for (int i = startBlock.getY() + offset; i <= playerPos.getY() - 1; i++) {
                     pos = new BlockPos(startBlock.getX(), i, startBlock.getZ());
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.DOWN) {
-                for (int i = startBlock.getY() - 1; i >= playerPos.getY() + 1; i--) {
+                for (int i = startBlock.getY() - offset; i >= playerPos.getY() + 1; i--) {
                     pos = new BlockPos(startBlock.getX(), i, startBlock.getZ());
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
@@ -109,9 +114,9 @@ public class BuildingModes {
         //***************************************************
         else if (mode == GadgetBuilding.ToolMode.VerticalWall) {
             if (sideHit == EnumFacing.UP) {
-                for (int y = 1; y <= range; y++) {
-                    for (int x = boundX * -1; x <= boundX; x++) {
-                        for (int z = boundZ * -1; z <= boundZ; z++) {
+                for (int y = offset; y <= range; y++) {
+                    for (int x = -boundX; x <= boundX; x++) {
+                        for (int z = -boundZ; z <= boundZ; z++) {
                             pos = new BlockPos(startBlock.getX() + x, startBlock.getY() + y, startBlock.getZ() + z);
                             if (isReplaceable(world, pos, setBlock)) {
                                 coordinates.add(pos);
@@ -120,9 +125,9 @@ public class BuildingModes {
                     }
                 }
             } else if (sideHit == EnumFacing.DOWN) {
-                for (int y = 1; y <= range; y++) {
-                    for (int x = boundX * -1; x <= boundX; x++) {
-                        for (int z = boundZ * -1; z <= boundZ; z++) {
+                for (int y = offset; y <= range; y++) {
+                    for (int x = -boundX; x <= boundX; x++) {
+                        for (int z = -boundZ; z <= boundZ; z++) {
                             pos = new BlockPos(startBlock.getX() + x, startBlock.getY() - y, startBlock.getZ() + z);
                             if (isReplaceable(world, pos, setBlock)) {
                                 coordinates.add(pos);
@@ -131,9 +136,9 @@ public class BuildingModes {
                     }
                 }
             } else {
-                for (int y = bound; y >= bound * -1; y--) {
-                    for (int x = boundX * -1; x <= boundX; x++) {
-                        for (int z = boundZ * -1; z <= boundZ; z++) {
+                for (int y = bound; y >= -bound; y--) {
+                    for (int x = -boundX; x <= boundX; x++) {
+                        for (int z = -boundZ; z <= boundZ; z++) {
                             pos = new BlockPos(startBlock.getX() + x, startBlock.getY() - y, startBlock.getZ() + z);
                             if (isReplaceable(world, pos, setBlock)) {
                                 coordinates.add(pos);
@@ -148,21 +153,21 @@ public class BuildingModes {
         //***************************************************
         else if (mode == GadgetBuilding.ToolMode.VerticalColumn) {
             if (sideHit == EnumFacing.UP) {
-                for (int y = 1; y <= range; y++) {
+                for (int y = offset; y <= range; y++) {
                     pos = new BlockPos(startBlock.getX(), startBlock.getY() + y, startBlock.getZ());
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.DOWN) {
-                for (int y = 1; y <= range; y++) {
+                for (int y = offset; y <= range; y++) {
                     pos = new BlockPos(startBlock.getX(), startBlock.getY() - y, startBlock.getZ());
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else {
-                for (int y = bound * -1; y <= bound; y++) {
+                for (int y = -bound; y <= bound; y++) {
                     pos = new BlockPos(startBlock.getX(), startBlock.getY() - y, startBlock.getZ());
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
@@ -178,28 +183,28 @@ public class BuildingModes {
                 sideHit = playerFacing.getOpposite();
             }
             if (sideHit == EnumFacing.NORTH) {
-                for (int z = 1; z <= range; z++) {
+                for (int z = offset; z <= range; z++) {
                     pos = new BlockPos(startBlock.getX(), startBlock.getY(), startBlock.getZ() + z);
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.SOUTH) {
-                for (int z = 1; z <= range; z++) {
+                for (int z = offset; z <= range; z++) {
                     pos = new BlockPos(startBlock.getX(), startBlock.getY(), startBlock.getZ() - z);
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.EAST) {
-                for (int x = 1; x <= range; x++) {
+                for (int x = offset; x <= range; x++) {
                     pos = new BlockPos(startBlock.getX() - x, startBlock.getY(), startBlock.getZ());
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.WEST) {
-                for (int x = 1; x <= range; x++) {
+                for (int x = offset; x <= range; x++) {
                     pos = new BlockPos(startBlock.getX() + x, startBlock.getY(), startBlock.getZ());
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
@@ -215,8 +220,8 @@ public class BuildingModes {
                 sideHit = playerFacing.getOpposite();
             }*/
             if (sideHit == EnumFacing.NORTH) {
-                for (int z = 1; z <= range; z++) {
-                    for (int x = bound * -1; x <= bound; x++) {
+                for (int z = offset; z <= range; z++) {
+                    for (int x = -bound; x <= bound; x++) {
                         pos = new BlockPos(startBlock.getX() + x, startBlock.getY(), startBlock.getZ() - z);
                         if (isReplaceable(world, pos, setBlock)) {
                             coordinates.add(pos);
@@ -224,8 +229,8 @@ public class BuildingModes {
                     }
                 }
             } else if (sideHit == EnumFacing.SOUTH) {
-                for (int z = 1; z <= range; z++) {
-                    for (int x = bound * -1; x <= bound; x++) {
+                for (int z = offset; z <= range; z++) {
+                    for (int x = -bound; x <= bound; x++) {
                         pos = new BlockPos(startBlock.getX() + x, startBlock.getY(), startBlock.getZ() + z);
                         if (isReplaceable(world, pos, setBlock)) {
                             coordinates.add(pos);
@@ -233,8 +238,8 @@ public class BuildingModes {
                     }
                 }
             } else if (sideHit == EnumFacing.EAST) {
-                for (int x = 1; x <= range; x++) {
-                    for (int z = bound * -1; z <= bound; z++) {
+                for (int x = offset; x <= range; x++) {
+                    for (int z = -bound; z <= bound; z++) {
                         pos = new BlockPos(startBlock.getX() + x, startBlock.getY(), startBlock.getZ() + z);
                         if (isReplaceable(world, pos, setBlock)) {
                             coordinates.add(pos);
@@ -242,8 +247,8 @@ public class BuildingModes {
                     }
                 }
             } else if (sideHit == EnumFacing.WEST) {
-                for (int x = 1; x <= range; x++) {
-                    for (int z = bound * -1; z <= bound; z++) {
+                for (int x = offset; x <= range; x++) {
+                    for (int z = -bound; z <= bound; z++) {
                         pos = new BlockPos(startBlock.getX() - x, startBlock.getY(), startBlock.getZ() + z);
                         if (isReplaceable(world, pos, setBlock)) {
                             coordinates.add(pos);
@@ -251,8 +256,8 @@ public class BuildingModes {
                     }
                 }
             } else {
-                for (int x = bound * -1; x <= bound; x++) {
-                    for (int z = bound * -1; z <= bound; z++) {
+                for (int x = -bound; x <= bound; x++) {
+                    for (int z = -bound; z <= bound; z++) {
                         pos = new BlockPos(startBlock.getX() - x, startBlock.getY(), startBlock.getZ() + z);
                         if (isReplaceable(world, pos, setBlock)) {
                             coordinates.add(pos);
@@ -269,52 +274,52 @@ public class BuildingModes {
                 sideHit = playerFacing.getOpposite();
             }
             if (sideHit == EnumFacing.NORTH) {
-                for (int z = 1; z <= range; z++) {
+                for (int z = offset; z <= range; z++) {
                     if (startBlock.getY() > player.posY + 1) {
                         pos = new BlockPos(startBlock.getX(), startBlock.getY() - z, startBlock.getZ() - z);
                     } else if (startBlock.getY() < player.posY - 2) {
-                        pos = new BlockPos(startBlock.getX(), startBlock.getY() + z, startBlock.getZ() - z + 1);
+                        pos = new BlockPos(startBlock.getX(), startBlock.getY() + z, startBlock.getZ() - z + offset);
                     } else {
-                        pos = new BlockPos(startBlock.getX(), startBlock.getY() + z, startBlock.getZ() + z - 1);
+                        pos = new BlockPos(startBlock.getX(), startBlock.getY() + z, startBlock.getZ() + z - offset);
                     }
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.SOUTH) {
-                for (int z = 1; z <= range; z++) {
+                for (int z = offset; z <= range; z++) {
                     if (startBlock.getY() > player.posY + 1) {
                         pos = new BlockPos(startBlock.getX(), startBlock.getY() - z, startBlock.getZ() + z);
                     } else if (startBlock.getY() < player.posY - 2) {
-                        pos = new BlockPos(startBlock.getX(), startBlock.getY() + z, startBlock.getZ() + z - 1);
+                        pos = new BlockPos(startBlock.getX(), startBlock.getY() + z, startBlock.getZ() + z - offset);
                     } else {
-                        pos = new BlockPos(startBlock.getX(), startBlock.getY() + z, startBlock.getZ() - z + 1);
+                        pos = new BlockPos(startBlock.getX(), startBlock.getY() + z, startBlock.getZ() - z + offset);
                     }
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.EAST) {
-                for (int x = 1; x <= range; x++) {
+                for (int x = offset; x <= range; x++) {
                     if (startBlock.getY() > player.posY + 1) {
                         pos = new BlockPos(startBlock.getX() + x, startBlock.getY() - x, startBlock.getZ());
                     } else if (startBlock.getY() < player.posY - 2) {
-                        pos = new BlockPos(startBlock.getX() + x - 1, startBlock.getY() + x, startBlock.getZ());
+                        pos = new BlockPos(startBlock.getX() + x - offset, startBlock.getY() + x, startBlock.getZ());
                     } else {
-                        pos = new BlockPos(startBlock.getX() - x + 1, startBlock.getY() + x, startBlock.getZ());
+                        pos = new BlockPos(startBlock.getX() - x + offset, startBlock.getY() + x, startBlock.getZ());
                     }
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
                     }
                 }
             } else if (sideHit == EnumFacing.WEST) {
-                for (int x = 1; x <= range; x++) {
+                for (int x = offset; x <= range; x++) {
                     if (startBlock.getY() > player.posY + 1) {
                         pos = new BlockPos(startBlock.getX() - x, startBlock.getY() - x, startBlock.getZ());
                     } else if (startBlock.getY() < player.posY - 2) {
-                        pos = new BlockPos(startBlock.getX() - x + 1, startBlock.getY() + x, startBlock.getZ());
+                        pos = new BlockPos(startBlock.getX() - x + offset, startBlock.getY() + x, startBlock.getZ());
                     } else {
-                        pos = new BlockPos(startBlock.getX() + x - 1, startBlock.getY() + x, startBlock.getZ());
+                        pos = new BlockPos(startBlock.getX() + x - offset, startBlock.getY() + x, startBlock.getZ());
                     }
                     if (isReplaceable(world, pos, setBlock)) {
                         coordinates.add(pos);
@@ -330,7 +335,7 @@ public class BuildingModes {
             for (int x = range * -7 / 5; x <= range * 7 / 5; x++) {
                 for (int z = range * -7 / 5; z <= range * 7 / 5; z++) {
                     if (x % (((range - 2) % 6) + 2) == 0 && z % (((range - 2) % 6) + 2) == 0) {
-                        pos = new BlockPos(startBlock.getX() + x, startBlock.getY() + 1, startBlock.getZ() + z);
+                        pos = new BlockPos(startBlock.getX() + x, startBlock.getY() + offset, startBlock.getZ() + z);
                         if (isReplaceable(world, pos, setBlock)) {
                             coordinates.add(pos);
                         }
@@ -348,11 +353,11 @@ public class BuildingModes {
             BlockPos locOffset;
             boolean fuzzyMode = GadgetGeneric.getFuzzy(tool);
             if (GadgetGeneric.getConnectedArea(tool)) {
-                addConnectedCoords(world, pos, startState, setBlock, sideHit, fuzzyMode, coordinates, new HashSet<>(),
+                addConnectedCoords(world, pos, onTop, startState, setBlock, sideHit, fuzzyMode, coordinates, new HashSet<>(),
                         (int) area.minX, (int) area.minY, (int) area.minZ, (int) area.maxX - 1, (int) area.maxY - 1, (int) area.maxZ - 1);
             } else {
                 for (BlockPos loc : BlockPos.getAllInBox((int) area.minX, (int) area.minY, (int) area.minZ, (int) area.maxX - 1, (int) area.maxY - 1, (int) area.maxZ - 1)) {
-                    locOffset = loc.offset(sideHit);
+                    locOffset = onTop ? loc.offset(sideHit) : loc;
                     if ((fuzzyMode ? !world.isAirBlock(loc) : world.getBlockState(loc) == startState) && isReplaceable(world, locOffset, setBlock))
                         coordinates.add(locOffset);
                 }
@@ -361,12 +366,12 @@ public class BuildingModes {
         return coordinates;
     }
 
-    private static void addConnectedCoords(World world, BlockPos loc, IBlockState state, IBlockState setBlock, EnumFacing sideHit,
+    private static void addConnectedCoords(World world, BlockPos loc, boolean onTop, IBlockState state, IBlockState setBlock, EnumFacing sideHit,
             boolean fuzzyMode, List<BlockPos> coords, Set<BlockPos> coordsSearched, int minX, int minY, int minZ, int maxX, int maxY, int maxZ) {
         if (coordsSearched.contains(loc) || loc.getX() < minX || loc.getY() < minY || loc.getZ() < minZ || loc.getX() > maxX || loc.getY() > maxY || loc.getZ() > maxZ)
             return;
 
-        BlockPos locOffset = loc.offset(sideHit);
+        BlockPos locOffset = onTop ? loc.offset(sideHit) : loc;
         if ((fuzzyMode ? world.isAirBlock(loc) : world.getBlockState(loc) != state) || !isReplaceable(world, locOffset, setBlock))
             return;
 
@@ -375,7 +380,7 @@ public class BuildingModes {
         for (int x = -1; x <= 1; x++) {
             for (int y = -1; y <= 1; y++) {
                 for (int z = -1; z <= 1; z++) {
-                    addConnectedCoords(world, loc.add(x, y, z), state, setBlock, sideHit, fuzzyMode, coords, coordsSearched, minX, minY, minZ, maxX, maxY, maxZ);
+                    addConnectedCoords(world, loc.add(x, y, z), onTop, state, setBlock, sideHit, fuzzyMode, coords, coordsSearched, minX, minY, minZ, maxX, maxY, maxZ);
                 }
             }
         }