+++ src/main/java/com/direwolf20/buildinggadgets/common/tools/ToolRenders.java
@@ -9,7 +9,6 @@ import com.direwolf20.buildinggadgets.common.items.capability.CapabilityProvider
 import com.direwolf20.buildinggadgets.common.items.gadgets.GadgetBuilding;
 import com.direwolf20.buildinggadgets.common.items.gadgets.GadgetCopyPaste;
 import com.direwolf20.buildinggadgets.common.items.gadgets.GadgetDestruction;
-import com.direwolf20.buildinggadgets.common.items.gadgets.GadgetExchanger;
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.collect.Multiset;
@@ -99,7 +98,7 @@ public class ToolRenders {
             GlStateManager.popMatrix();
         }
 
-        RayTraceResult lookingAt = VectorTools.getLookingAt(player);
+        RayTraceResult lookingAt = VectorTools.getLookingAt(player, heldItem);
         IBlockState state = Blocks.AIR.getDefaultState();
         List<BlockPos> coordinates = getAnchor(stack);
         if (lookingAt != null || coordinates.size() > 0) {
@@ -214,17 +213,14 @@ public class ToolRenders {
     }
 
     public static void renderExchangerOverlay(RenderWorldLastEvent evt, EntityPlayer player, ItemStack stack) {
-        ItemStack heldItem = GadgetExchanger.getGadget(player);
-        if (heldItem.isEmpty()) return;
-
         //Calculate the players current position, which is needed later
         double doubleX = player.lastTickPosX + (player.posX - player.lastTickPosX) * evt.getPartialTicks();
         double doubleY = player.lastTickPosY + (player.posY - player.lastTickPosY) * evt.getPartialTicks();
         double doubleZ = player.lastTickPosZ + (player.posZ - player.lastTickPosZ) * evt.getPartialTicks();
 
         BlockRendererDispatcher dispatcher = Minecraft.getMinecraft().getBlockRendererDispatcher();
-        Integer dim = GadgetUtils.getDIMFromNBT(heldItem, "boundTE");
-        BlockPos pos = GadgetUtils.getPOSFromNBT(heldItem, "boundTE");
+        Integer dim = GadgetUtils.getDIMFromNBT(stack, "boundTE");
+        BlockPos pos = GadgetUtils.getPOSFromNBT(stack, "boundTE");
 
         if (dim != null && pos != null) {
             GlStateManager.pushMatrix();//Push matrix again just because
@@ -242,7 +238,7 @@ public class ToolRenders {
             GlStateManager.popMatrix();
         }
 
-        RayTraceResult lookingAt = VectorTools.getLookingAt(player);
+        RayTraceResult lookingAt = VectorTools.getLookingAt(player, stack);
         IBlockState state = Blocks.AIR.getDefaultState();
         List<BlockPos> coordinates = getAnchor(stack);
         if (lookingAt != null || coordinates.size() > 0) {
@@ -252,9 +248,7 @@ public class ToolRenders {
                 startBlock = world.getBlockState(lookingAt.getBlockPos());
             }
             if (startBlock != ModBlocks.effectBlock.getDefaultState()) {
-
-
-                IBlockState renderBlockState = getToolBlock(heldItem);
+                IBlockState renderBlockState = getToolBlock(stack);
                 Minecraft mc = Minecraft.getMinecraft();
                 mc.renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
                 if (renderBlockState == Blocks.AIR.getDefaultState()) {//Don't render anything if there is no block selected (Air)
@@ -346,9 +340,9 @@ public class ToolRenders {
                     GL14.glBlendColor(1F, 1F, 1F, 0.55f); //Set the alpha of the blocks we are rendering
                     hasBlocks--;
                     if (stack.hasCapability(CapabilityEnergy.ENERGY, null)) {
-                        hasEnergy -= ModItems.gadgetExchanger.getEnergyCost(heldItem);
+                        hasEnergy -= ModItems.gadgetExchanger.getEnergyCost(stack);
                     } else {
-                        hasEnergy -= ModItems.gadgetExchanger.getDamageCost(heldItem);
+                        hasEnergy -= ModItems.gadgetExchanger.getDamageCost(stack);
                     }
                     if (hasBlocks < 0 || hasEnergy < 0) {
                         dispatcher.renderBlockBrightness(Blocks.STAINED_GLASS.getDefaultState().withProperty(COLOR, EnumDyeColor.RED), 1f);
@@ -370,28 +364,24 @@ public class ToolRenders {
     }
 
     public static void renderDestructionOverlay(RenderWorldLastEvent evt, EntityPlayer player, ItemStack stack) {
-        RayTraceResult lookingAt = VectorTools.getLookingAt(player);
+        RayTraceResult lookingAt = VectorTools.getLookingAt(player, stack);
         if (lookingAt == null && GadgetDestruction.getAnchor(stack) == null) return;
         World world = player.world;
         BlockPos startBlock = (GadgetDestruction.getAnchor(stack) == null) ? lookingAt.getBlockPos() : GadgetDestruction.getAnchor(stack);
         EnumFacing facing = (GadgetDestruction.getAnchorSide(stack) == null) ? lookingAt.sideHit : GadgetDestruction.getAnchorSide(stack);
         if (startBlock == ModBlocks.effectBlock.getDefaultState()) return;
 
-        ItemStack heldItem = GadgetDestruction.getGadget(player);
-        if (heldItem.isEmpty()) return;
-
-        if (!GadgetDestruction.getOverlay(heldItem)) return;
-
+        if (!GadgetDestruction.getOverlay(stack)) return;
         GlStateManager.pushMatrix();
         double doubleX = player.lastTickPosX + (player.posX - player.lastTickPosX) * evt.getPartialTicks();
         double doubleY = player.lastTickPosY + (player.posY - player.lastTickPosY) * evt.getPartialTicks();
         double doubleZ = player.lastTickPosZ + (player.posZ - player.lastTickPosZ) * evt.getPartialTicks();
         GlStateManager.translate(-doubleX, -doubleY, -doubleZ);
         try {
-            GlStateManager.callList(cacheDestructionOverlay.get(new ImmutableTriple<UniqueItemStack, BlockPos, Integer>(new UniqueItemStack(heldItem), startBlock, facing.ordinal()), () -> {
+            GlStateManager.callList(cacheDestructionOverlay.get(new ImmutableTriple<UniqueItemStack, BlockPos, Integer>(new UniqueItemStack(stack), startBlock, facing.ordinal()), () -> {
                 int displayList = GLAllocation.generateDisplayLists(1);
                 GlStateManager.glNewList(displayList, GL11.GL_COMPILE);
-                renderDestructionOverlay(player, world, startBlock, facing, heldItem);
+                renderDestructionOverlay(player, world, startBlock, facing, stack);
                 GlStateManager.glEndList();
                 return displayList;
             }));
@@ -506,7 +496,7 @@ public class ToolRenders {
             //First check if we have an anchor, if not check if we're looking at a block, if not, exit
             BlockPos startPos = GadgetCopyPaste.getAnchor(stack);
             if (startPos == null) {
-                startPos = VectorTools.getPosLookingAt(player);
+                startPos = VectorTools.getPosLookingAt(player, stack);
                 if (startPos == null) return;
                 startPos = startPos.up(GadgetCopyPaste.getY(stack));
                 startPos = startPos.east(GadgetCopyPaste.getX(stack));