+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format of this document is based on [Keep a Changelog](https://keepachangelo
 ## [Unreleased] 
 ### Added
 - Allowed gadgets to bind Applied Energistics 2 networks as a remote inventories.
+- Added gadget modes for:
+    - [all gadgets] raytracing fluids
+    - [building gadget] placing in or on blocks
 - More German localisation
 
 ### Fixed